+++ Solutions/Corvus.Testing.SpecFlow.Specs/Containers/ContainerTestContext.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.SpecFlow.Extensions.Specs.Containers
+namespace Corvus.Testing.SpecFlow.Specs.Containers
 {
     using System;
     using System.Collections.Generic;
