+++ Solutions/Corvus.Testing.SpecFlow/Corvus/Testing/SpecFlow/ContainerBeforeFeatureOrder.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.SpecFlow.Extensions
+namespace Corvus.Testing.SpecFlow
 {
     using System;
     using TechTalk.SpecFlow;
