+++ Solutions/Corvus.Testing.SpecFlow/Corvus/Testing/SpecFlow/ContainerBindings.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.SpecFlow.Extensions
+namespace Corvus.Testing.SpecFlow
 {
     using System;
     using Microsoft.Extensions.DependencyInjection;
