+++ Solutions/Corvus.Testing.SpecFlow/Corvus/Testing/SpecFlow/Internal/ChildObjectValueRetriever.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Endjin Limited. All rights reserved.
 // </copyright>
 
-namespace Corvus.SpecFlow.Extensions.Internal
+namespace Corvus.Testing.SpecFlow.Internal
 {
     using System;
     using System.Collections.Generic;
